pyproject.toml
[tool.poetry]
name = "ai_engineer"
version = "0.1.0"
description = "A python lib for integrating with various AI models, starting with services from openai, to co-engineer with you on a project level."
authors = ["Stefan du Plessis <67509094+stefanlduplessis@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "ai_engineer", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "^0"
anyio = "^4"
certifi = "^2024"
colorama = "^0"
distro = "^1"
h11 = "^0"
httpcore = "^1"
httpx = "^0"
idna = "^3"
openai = "^1"
pydantic = "^2"
pydantic-core = "^2"
sniffio = "^1"
tqdm = "^4"
typing-extensions = "^4"
python-dotenv = "^1.0.1"
typer = {extras = ["all"], version = "^0.13.1"}


[tool.poetry.group.dev.dependencies]
pydevd = "^3.1.0"
isort = "^5.13.2"
pylint = "^3.2.6"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ai_engineer_cli = "ai_engineer.main:cli"
